// frontend/src/lib/api/generated/model/billTrackingItemDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ระบบรายงานค่าสาธารณูปโภค กรมประมง API
 * OpenAPI spec version: 2.0.0
 */

export interface BillTrackingItemDTO {
  amount: number;
  billId?: string;
  departmentId: string;
  departmentName: string;
  id: string;
  isExpected: boolean;
  provider: string;
  serviceNumber: string;
  status: string;
  utilityType: string;
}