// frontend/src/lib/api/generated/model/billTrackingSummaryDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ระบบรายงานค่าสาธารณูปโภค กรมประมง API
 * OpenAPI spec version: 2.0.0
 */

export interface BillTrackingSummaryDTO {
  notReceived: number;
  paid: number;
  pending: number;
  total: number;
  unrecorded: number;
}